+++ src/mousetrap/engine/__init__.py
@@ -1,0 +1,3 @@
+from .engine import Engine
+from .bus import Bus, Event
+from .component import Component
